Doing a little cleanup
